Implement algorithm to solve Boggle word game
Boggle is a word game played on an 4-by-4 grid of tiles,
where each tile contains one letter in the alphabet.
The goal is to find all words in the dictionary that can
be made by following a path of adjacent tiles (with no tile repeated),
where two tiles are adjacent if they are horizontal, vertical,
or diagonal neighbors.